+++ R/internal-lin_alg.R
@@ -775,12 +775,14 @@ is_conv_to_asymptotic <- function(ipm, tol = 1e-10) {
 # Helper function to handle when burn_in == 0
 #' @noRd
 
-.thin_stoch_lambda <- function(lambdas, burn_ind) {
+.thin_stoch_lambda <- function(lambdas, burn_ind, log) {
+
+  if(log) lambdas <- log(lambdas)
 
   if(length(burn_ind > 0)) {
-    out <- mean(log(lambdas[-c(burn_ind)]))
+    out <- mean(lambdas[-c(burn_ind)])
   } else {
-    out <- mean(log(lambdas))
+    out <- mean(lambdas)
   }
 
   return(out)