+++ tests/testthat/test-general_dd_stoch_param.R
@@ -218,6 +218,7 @@ hand_pop_sizes <- lapply(pop_holder,
 
 ipmr_lams <- lambda(gen_dd_stoch_co, type_lambda = "all") %>%
   as.vector()
+
 hand_lams <- hand_pop_sizes[2:51] / hand_pop_sizes[1:50]
 
 test_that("models compute population sizes correctly", {
@@ -227,5 +228,5 @@ test_that("models compute population sizes correctly", {
 
 })
 
-# Needs test for par_setarchical models
+# Needs test for par_set models